+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - registry/storage: Cache repository tags during the garbage collection mark phase
 - registry/storage: Garbage collection provides an estimate of how much storage will be removed
 - registry/storage: Upgrade the GCS SDK to v1.2.1
+- registry/api/v2: Return not found error when getting a manifest by tag with a broken link
 
 ## v2.8.1-gitlab